+++ client/components/more-on/main.js
@@ -34,8 +34,8 @@ var $ = function(selector) {
 module.exports.init = function(flags){
 	var fetchPromises = [];
 
-	$('.js-more-on').forEach(function(el) {
-		fetchPromises.push(fetch('/more-on/' + el.getAttribute('data-article-id'))
+	$('.js-more-on-inline').forEach(function(el) {
+		fetchPromises.push(fetch('/more-on/' + document.querySelector('.article').getAttribute('data-capi-id') + '?count=1&view=inline')
 			.then(fetchres.text)
 			.then(function(resp) {
 				el.innerHTML = resp;
@@ -45,31 +45,18 @@ module.exports.init = function(flags){
 			}));
 	});
 
-	var inlineRelatedAnchor = document.querySelector('.js-more-on-inline');
-	if (inlineRelatedAnchor) {
-		fetchPromises.push(fetch('/' + document.querySelector('.article').getAttribute('data-capi-id') + '/related?count=1&view=inline')
+	$('.js-more-on').forEach(function(el) {
+		fetchPromises.push(fetch('/more-on/' + document.querySelector('.article').getAttribute('data-capi-id') + '?count=4')
 			.then(fetchres.text)
 			.then(function(resp) {
-				inlineRelatedAnchor.innerHTML = resp;
-				oDate.init(inlineRelatedAnchor);
-			}, function() {
-				inlineRelatedAnchor.parentNode.removeChild(inlineRelatedAnchor);
-			}));
-	}
-
-	var relatedAnchor = document.querySelector('.js-related');
-	if (relatedAnchor) {
-		fetchPromises.push(fetch('/' + document.querySelector('.article').getAttribute('data-capi-id') + '/related?count=4')
-			.then(fetchres.text)
-			.then(function(resp) {
-				relatedAnchor.innerHTML = resp;
+				el.innerHTML = resp;
 				oDate.init(relatedAnchor);
 			}, function() {
-				relatedAnchor.parentNode.removeChild(relatedAnchor);
+				el.parentNode.removeChild(el);
 			}));
-	}
+	});
 
-	$('.js-on-this-topic').forEach(function(el) {
+	$('.js-more-on-topic').forEach(function(el) {
 		fetchPromises.push(fetch('/more-on/' + el.getAttribute('data-metadata-field') + '/' + el.getAttribute('data-article-id'))
 			.then(fetchres.text)
 			.then(function(resp) {
