+++ views/layout.html
@@ -75,9 +75,9 @@
 {{#if @root.flags.articleRelatedContent.isSwitchedOn}}
 <div class="article__related">
 	<div class="o-grid-row">
-		<div class="js-related" data-o-grid-colspan="12 L4"></div>
-		<div class='more-on js-more-on' data-article-id='{{id}}' data-o-grid-colspan="12 L4"></div>
-		<div class="on-this-topic js-on-this-topic" data-metadata-field="primarySection" data-article-id="{{id}}" data-trackable="more-on-section" data-o-grid-colspan="12 L4"></div>
+		<div class="js-more-on" data-o-grid-colspan="12 L4"></div>
+		<div class='js-more-on-topic' data-metadata-field="primaryTheme" data-article-id='{{id}}' data-o-grid-colspan="12 L4" data-trackable="more-on-theme" ></div>
+		<div class="js-more-on-topic" data-metadata-field="primarySection" data-article-id="{{id}}" data-o-grid-colspan="12 L4" data-trackable="more-on-section" ></div>
 	</div>
 </div>
 {{/if}}
