--- views/related.html
@@ -1,17 +1,0 @@
-<div class="related{{#if isInline}} ng-pull-out ng-inline-element{{/if}}" data-trackable="related{{#if isInline}}-inline{{/if}}">
-	<h3 class="related__title">Related {{#ifEquals articles.length 1}}story{{else}}stories{{/ifEquals}}</h3>
-	<ul class="related__articles ng-list-reset" data-trackable="articles">
-		{{#each articles}}
-			<li class="related__article">
-				<a class="related__article__link ng-title-link" href="/{{id}}" data-trackable="title">{{{title}}}</a>
-				<time
-					class="related__article__timestamp ng-meta o-date"
-					data-o-component="o-date"
-					datetime="{{#dateformat}}{{publishedDate}}{{/dateformat}}"
-					data-o-date-js>
-					{{#dateformat "dddd, d mmmm, yyyy"}}{{publishedDate}}{{/dateformat}}
-				</time>
-			</li>
-		{{/each}}
-	</ul>
-</div>
